Replace math punctuation transform
Enable and add a new math_punctuation transform, remove the old implementation, and update config. book/_ext/math_punctuation.py implements InlineMathPunctuationIntoText (priority 700) which moves sentence-ending punctuation immediately following inline math into the math content as \text{...}, replacing the math node and trimming/removing the following text node. The previous math_punctuation_into_text.py was deleted and book/_config.yml was updated to reference math_punctuation instead of the old name.
